Add database integration and logging for workout plans
- Implemented a new repository for managing workout plans in the database.
- Updated the database client structure to expose the DB field.
- Enhanced the service to create a database client and save workout plans.
- Added a Justfile for easier service execution.
- Updated README to reflect future enhancements for logging previous workouts.
