+++ justfile
@@ -1,0 +1,4 @@
+run-service:
+    set -a
+    source .env
+    go run cmd/ai-trainer/main.go
